common: De-virtualize VDF
Also, the random number generators in the VDF class is refactored.

This is the first step to increase performance by removing
the virtual function dispatch. In this modification, the templated VDF
class provide the public interface and use the CRTP mechanism to dispatch
the function calls to the appropriate concrete VDF object at compile time.

The next step is to implement VDFVariant which will hold the all possible
VDF concrete types.