pic_1d: Fix critical bug of ColdSpecies moment collection when particle parallelism is enabled
This fixes a critical bug whereby all ColdSpecies moments were not accounted for
when the particle parallelism is enabled. This was caused by the removal of ColdSpecies pass
when first entering Domain::advance, where passing a grid quantity automatically
distributes it to all worker threads. This design is awkward as well as flawed.

This patch fixes this issue by redesigning the ColdSpecies sharing mechanism on a multi-
threaded environment, where ColdSpecies moments are evenly split among the worker threads,
thereby resulting in symmetry between particle and cold species. This also removes the
modification of PartSpecies::Nc from within Master/WorkerDelegate's.