hybrid_1d+pic_1d+rel_pic_1d: Fix an issue where equilibrium values were not split when using delta-f scheme
After switching to dividing up the cold and particle species in multi-threading parallelism,
splitting the equilibrium values before adding them to the moments of delta-f particles
was not taken care of.

In addition, this patch also fixes a separate issue where the equilibrium values should have not
been added to the ghost cells.