pic_1d: Snapshot ghost cells of grid quantities
This patch is in preparation for the fix of moment collection at the left physical boundary
when the particle reflection boundary condition is chosen.

Saving the values at the ghost cells disables the change of the number of subdomains
between the snapshot-and-resumes. To prevent this, the number_of_subdomains parameter
is also part of the parameter signature.