pic_1d: Fix moment gather on the use of reflecting particle boundary condition
Previously, when the reflecting particle boundary condition is chosen, the moments
at the left ghost grid points at the leftmost subdomain are accumulated to the first
grid point. Likewise, the moments at the right ghost grid points are accumulated to
the last grid point. This resulted in unbalance of moments at the first physical grid
point and the last physical point, because particles are confined between 0 and Nx:

              0     1     2     .......    Nx-1   Nx
              |-----|-----|-----|-----|-----|-----|

When for example the CIC shape is used, the moments at the first (0th) grid point only come
from the particles in the first cell, which are shared with the second grid point.
On the other hand, the moments at the last (Nx-1th) grid point are contributed from the
particles in the last two cells. Therefore, the previous scheme resulted in half-the-average
moments at the first grid point and three-halves-the-average moments at the last grid point.

In this patch, the density collected at the Nx grid point is added to the first grid point,
whereas all other higher moments at the first grid point is doubled of the accumulated values.
The accumulation of density is to ensure the conservation of the total number of particles.