docs: Modify grid strategy of pic_scheme.nb
The electric field and the grid quantities aligned with the electric field are put into
the half-integer grid points, whereas the magnetic field are put into the full-integer
grid points. The bounding box of the simulation particles coincide with the full-integer
grid points.

This change is to simplify the moment gathering when the particle reflection option is
used.