libPIC: Use the physical to marker particle PSD ratio in calculating weight in the full-f scheme
This patch is in preparation for PSD refreshment. When the full-f scheme is chosen, simulation
particles' PSD is initialized with appropriate physical PSD and marker PSD with a weight determined
by the ratio of the physical to marker PSDs.

This change will lead to a bug in all-moment collection when marker_temp_ratio is set
to a value other than one because of the weight calculation.